refactor(rez): move materialize_wal_tail to the writer that owns WAL rows
It builds a segment out of `WalCell`s, which the v3 writer defines and
encodes, so it belongs beside them rather than in the reader that
happened to be its first caller. The reader keeps calling it; the writer
is about to, for sealing, and would otherwise have to reach into the
read path to seal a segment.
